requirements.txt
Pinata-python
web3
python-dotenv
wave
openai
requests
playsound
pydub
sounddevice
numpy
nltk
speechrecognition
#scipy
pyserial
#mutagen
#regex_spm
toml
pygame
keyboard
#eth-tester
#py-evm
oz-defender
torch
openai-whisper
whisper-mic
langchain
elevenlabs
simpleaichat